autenticação e jwt

+++ server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { prisma } from "../lib/prisma";
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+
+    // permite aplicar verificação do JWT em todas as rotas
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+
+    app.get('/memories', async (request) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 createdAt: 'asc',
 
@@ -20,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -32,6 +41,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             }
         })
+
+        if (!memory.isPublic &&  memory.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
 
         return memory
     })
@@ -50,14 +63,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: 'e617409c-33fe-493e-ba29-688e1e7ed592'
+                userId: request.user.sub
             }
         })
 
         return memory
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -72,7 +85,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = boodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId !== request.user.sub) return reply.status(401).send()
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -86,12 +107,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async (request, response) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId !== request.user.sub) return reply.status(401).send()
 
         await prisma.memory.delete({
             where: {
@@ -99,6 +128,6 @@ export async function memoriesRoutes(app: FastifyInstance) {
             }
         })
 
-        return response.send({ message: "Memory deleted" })
+        return reply.status(200).send({ message: "Memory deleted" })
     })
 }
